- client: change some unsigned int to size_t in our versions         of NVIDIA APIs.  This apparently caused crashes         (in app, not client, which I don't understand) for Einstein@Home.         From Steffen Moller.
svn path=/branches/boinc_core_release_7_0a/; revision=25568

+++ lib/coproc.h
@@ -263,7 +263,7 @@ struct COPROC {
 struct CUDA_DEVICE_PROP {
   char  name[256];
   int   deviceHandle;
-  unsigned int totalGlobalMem;
+  size_t totalGlobalMem;
     // not used on the server; dtotalGlobalMem is used instead
     // (since some boards have >= 4GB)
   int   sharedMemPerBlock;
